Added many, many things
Added an idle timer that automatically sends Jarvis into a "suspended
state" after 60 seconds of inactivity.

Edited the timer code to ensure that the counter is set to zero if the
log is open.

Added StartScript to control the "suspended state" scene, and allow for
easy manual and automatic switching between it and the main scene.

Added the updated sprites to Jarvis' animations.